// Richbourn-Dashboard/src/app/core/models/location.model.ts
export interface LocationModel {
    cities:any
    Location: string;
    Series: string;

    id: any;
    SlNo: any;
    Description: String;
    City: String;
    CityID:number;
    LocationType: String;
    Address: String;
    Email : String;
    Country : String;
    Zip: String;
    State: String;
    Telephone : String;
    Contact: String;
}